flow/requirements.txt
# Copyright 2025 Lunch Pail Labs, LLC
# Licensed under the Apache License, Version 2.0

# FastMCP and core dependencies
mcp>=0.1.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.8.0
python-dotenv>=1.0.0
unkey-py>=0.3.0
langgraph==0.2.40

# Pipecat for AI bot functionality
pipecat-ai[daily,openai,local-smart-turn-v3]>=0.0.94

# Image processing for sprite animation
Pillow>=10.0.0

# Email service
resend>=2.0.0

# Field-level encryption
cryptography>=42.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
mypy>=1.5.0
ruff>=0.3.0
pre-commit>=3.5.0
